Restructure Scene Tree menu
